requirements.txt
azure-functions
requests
azure-storage-blob 